Fix mapping of league to sport for hockey
The URL for fantasy hockey is now `hockey.fantasysports.yahoo.com/hockey` instead of `nhl.fantasysports.yahoo.com/hockey`, so I've adjusted the mapping to clarify this. I have not checked if other sports require this change as well, and thus have just modified the mapping function rather than refactor it out (which should be done if this change is across all sports).